[REFACTORING] MapperConfig 분리 / 메시지 Pub
